+++ tsconfig.json
@@ -4,7 +4,6 @@
     "target": "ES2015",
     "module": "umd",
     "declaration": true,
-    // "declarationDir": "./dist/types",
     "outDir": "./dist",
     "strict": true,
     "typeRoots": [
